Finish implementation and add javadocs
Had some small errors and unfinished docs

Should be cleaned up and documentation thrown in for future reference
