+++ src/main/java/commandhandling/DeadlineHandler.java
@@ -1,20 +1,28 @@
 package commandhandling;
 
 import commandhandling.argsorting.ArgSorter;
+import exceptions.syntax.KniazInvalidArgsException;
 import main.KniazSession;
 import task.Deadline;
 import task.Task;
 
 /**
- * Encapsulation of an abstract class that handles the logic and input validation
- * of Deadline(see task.Deadline) creation
- * Includes handling of arguments into Deadline construction
+ * Encapsulates the execution of a deadline commmand, creating a new Deadline.
  */
 public class DeadlineHandler implements CommandHandler {
 
+    // the expected order of arguments to this command
     private static final String[] ARG_ORDER = new String[]{"","/by"};
+
+    /**
+     * Executes the command to create a new Deadline
+     * @param session the linked KniazSession that this command is to execute in
+     * @param args    the arguments to this command
+     * @return the user-facing string representation of the created Deadline
+     * @throws  KniazInvalidArgsException when the arguments are invalid, such as not having a "/by {TIME}" argument
+     */
     @Override
-    public String handle(KniazSession session, String[] args) {
+    public String handle(KniazSession session, String[] args) throws KniazInvalidArgsException {
 
 
         String[] sortedArgs = ArgSorter.sortArgsByStarting(args,ARG_ORDER);
