+++ src/main/java/commandhandling/DeleteHandler.java
@@ -10,16 +10,24 @@ import task.Task;
 import ui.inputparser.InstructionType;
 
 import java.util.List;
+
 /**
- * Abstract class containing logic of deleting a task from a list
- * Also handles input validation
+ * Handles the delete command, by deleting the specified task.
  */
 public class DeleteHandler implements CommandHandler {
 
+    // the expected order of arguments to this command
     private static final String[] ARG_ORDER = new String[]{""};
 
+    /**
+     * Executes the deletion of a Task from the KniazSession's Tasklist.
+     * @param session the linked KniazSession that this command is to execute in
+     * @param args the arguments to this command
+     * @return the user-facing string representation of the deleted task
+     * @throws KniazInvalidArgsException when the argument is invalid, like the index being out of bounds
+     */
     @Override
-    public String handle(KniazSession session, String[] args) {
+    public String handle(KniazSession session, String[] args) throws KniazInvalidArgsException{
 
         String[] sortedArgs = ArgSorter.sortArgsByStarting(args,ARG_ORDER);
         // appears redundant but acts as another gatekeeper to make sure the arg syntax is right
