+++ src/main/java/commandhandling/ToDoHandler.java
@@ -1,6 +1,8 @@
 package commandhandling;
 
+import commandhandling.argsorting.ArgSorter;
 import exceptions.KniazRuntimeException;
+import exceptions.syntax.KniazInvalidArgsException;
 import main.KniazSession;
 import task.Task;
 import storage.TaskList;
@@ -9,15 +11,24 @@ import task.ToDo;
 import java.util.List;
 
 /**
- * Encapsulation of an abstract class that handles the logic and input validation
- * of ToDo(see task.ToDo) creation
- * Includes handling of arguments into ToDo construction
+ * Handles "todo" command by instantiating a new "todo".
  */
 public  class ToDoHandler implements CommandHandler {
-    @Override
-    public String handle(KniazSession session, String[] args) {
 
-        String toDoName = args[0];
+
+    private static final String[] ARG_ORDER = new String[]{""};
+    /**
+     * Handles todo command by making a new ToDo
+     * @param session the linked KniazSession that this command is to execute in
+     * @param args the arguments to this command
+     * @return the user-facing string representation of the newly made Task
+     * @throws KniazInvalidArgsException
+     */
+    @Override
+    public String handle(KniazSession session, String[] args) throws KniazInvalidArgsException {
+
+        String[] sortedArgs = ArgSorter.sortArgsByStarting(args,ARG_ORDER);
+        String toDoName = sortedArgs[0];
         Task taskToAdd = new ToDo(toDoName);
 
         session.getTaskList().add(taskToAdd);
