+++ src/main/java/commandhandling/UnmarkHandler.java
@@ -10,15 +10,22 @@ import task.Task;
 import java.util.List;
 
 /**
- * Abstract class containing logic of marking a task as not done.
+ * Handles the unmark command by un-marking specified task
  */
 public  class UnmarkHandler implements CommandHandler {
 
 
     private static final String[] ARG_ORDER = new String[]{""};
 
+    /**
+     * Handles unmark command by un-marking specified task
+     * @param session the linked KniazSession that this command is to execute in
+     * @param args the arguments to this command
+     * @return the user-facings tring representation of the unmarked task
+     * @throws KniazInvalidArgsException when there is a problem with the arguments, like index being out of bounds
+     */
     @Override
-    public String handle(KniazSession session, String[] args) {
+    public String handle(KniazSession session, String[] args) throws KniazInvalidArgsException{
 
         String[] sortedArgs = ArgSorter.sortArgsByStarting(args,ARG_ORDER);
         // appears redundant but acts as another gatekeeper to make sure the arg syntax is right
